Added protection by creator for books Added whitelist validation settings Updated validation settings

+++ src/features/books/services/books.service.ts
@@ -1,4 +1,4 @@
-import {Injectable} from '@nestjs/common';
+import {HttpException, HttpStatus, Injectable} from '@nestjs/common';
 
 import {InjectModel} from '@nestjs/mongoose';
 import {Model} from 'mongoose';
@@ -13,8 +13,11 @@ export class BooksService {
         @InjectModel(Book.name) private bookModel: Model<Book>
     ) { }
 
-    async create(createBookDto: CreateBookDto): Promise<Book> {
-        const createdBook = new this.bookModel(createBookDto);
+    async create(createBookDto: CreateBookDto, userName: string): Promise<Book> {
+        const createdBook = new this.bookModel({
+            ...createBookDto,
+            creator: userName,
+        });
 
         return createdBook.save();
     }
@@ -35,8 +38,18 @@ export class BooksService {
         return this.bookModel.findById(id).exec();
     }
 
-    async update(updateBookDto: UpdateBookDto): Promise<Book> {
+    async update(updateBookDto: UpdateBookDto, currentUserName: string): Promise<Book> {
         const {id, ...bookObject} = updateBookDto;
+
+        const book = await this.findOne(id);
+
+        if (book.creator !== currentUserName) {
+            throw new HttpException('Only the book creator can edit a book', HttpStatus.UNAUTHORIZED);
+        }
+
+        if (updateBookDto.creator !== currentUserName) {
+            throw new HttpException('Book creator reassignment is forbidden', HttpStatus.FORBIDDEN);
+        }
 
         const updatedBook = this.bookModel.findByIdAndUpdate(
             id,
@@ -47,7 +60,13 @@ export class BooksService {
         return updatedBook.exec();
     }
 
-    async remove(id: string): Promise<void> {
+    async remove(id: string, currentUserName: string): Promise<void> {
+        const book = await this.findOne(id);
+
+        if (book.creator !== currentUserName) {
+            throw new HttpException('Only the book creator can edit a book', HttpStatus.UNAUTHORIZED);
+        }
+
         await this.bookModel.findByIdAndDelete(id).exec();
 
         return;
